fix(tutorial): minor improvements on tutorial applications (#409)
* fix: Use own position when logging at which point the route was changed
* fix: separate event-tx and message-rx functionality
* fix: rather print connections than nodes because that's the more critical information
* fix: remove redundant information
* refactor: use precise wording, as the route switch may still fail
