+++ app/tutorials/traffic-light-communication/src/main/java/org/eclipse/mosaic/app/tutorial/EventSendingApp.java
@@ -16,18 +16,10 @@
 package org.eclipse.mosaic.app.tutorial;
 
 import org.eclipse.mosaic.app.tutorial.message.IntraVehicleMsg;
-import org.eclipse.mosaic.fed.application.ambassador.simulation.communication.AdHocModuleConfiguration;
-import org.eclipse.mosaic.fed.application.ambassador.simulation.communication.CamBuilder;
-import org.eclipse.mosaic.fed.application.ambassador.simulation.communication.ReceivedAcknowledgement;
-import org.eclipse.mosaic.fed.application.ambassador.simulation.communication.ReceivedV2xMessage;
 import org.eclipse.mosaic.fed.application.app.AbstractApplication;
 import org.eclipse.mosaic.fed.application.app.api.Application;
-import org.eclipse.mosaic.fed.application.app.api.CommunicationApplication;
 import org.eclipse.mosaic.fed.application.app.api.VehicleApplication;
 import org.eclipse.mosaic.fed.application.app.api.os.VehicleOperatingSystem;
-import org.eclipse.mosaic.interactions.communication.V2xMessageTransmission;
-import org.eclipse.mosaic.lib.enums.AdHocChannel;
-import org.eclipse.mosaic.lib.enums.SensorType;
 import org.eclipse.mosaic.lib.objects.vehicle.VehicleData;
 import org.eclipse.mosaic.lib.util.scheduling.Event;
 
@@ -35,7 +27,7 @@ import java.util.List;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
-public class EventSendingApp extends AbstractApplication<VehicleOperatingSystem> implements VehicleApplication, CommunicationApplication {
+public class EventSendingApp extends AbstractApplication<VehicleOperatingSystem> implements VehicleApplication {
     /**
      * Used for choosing a RAND id for the message that is sent intra-vehicle.
      */
@@ -44,23 +36,12 @@ public class EventSendingApp extends AbstractApplication<VehicleOperatingSystem>
     @Override
     public void onStartup() {
         getLog().infoSimTime(this, "Initialize application");
-        getOs().getAdHocModule().enable(new AdHocModuleConfiguration()
-                .addRadio()
-                .channel(AdHocChannel.CCH)
-                .power(50)
-                .create());
-        getLog().infoSimTime(this, "Activated AdHoc Module");
     }
 
     @Override
     public void onVehicleUpdated(@Nullable VehicleData previousVehicleData, @Nonnull VehicleData updatedVehicleData) {
         final List<? extends Application> applications = getOs().getApplications();
         final IntraVehicleMsg message = new IntraVehicleMsg(getOs().getId(), getRandom().nextInt(0, MAX_ID));
-
-        // Example usage for how to detect sensor readings
-        if (getOs().getStateOfEnvironmentSensor(SensorType.OBSTACLE) > 0) {
-            getLog().infoSimTime(this, "Reading sensor");
-        }
 
         for (Application application : applications) {
             final Event event = new Event(getOs().getSimulationTime() + 10, application, message);
@@ -69,30 +50,12 @@ public class EventSendingApp extends AbstractApplication<VehicleOperatingSystem>
     }
 
     @Override
-    public void onMessageReceived(ReceivedV2xMessage receivedV2xMessage) {
-        getLog().infoSimTime(this, "Received V2X Message from {}", receivedV2xMessage.getMessage().getRouting().getSource().getSourceName());
-    }
-
-    @Override
     public void processEvent(Event event) throws Exception {
-        getLog().infoSimTime(this, "Received event: {}", getOs().getSimulationTimeMs(), event.getResourceClassSimpleName());
+        getLog().infoSimTime(this, "Received event: {}", event.getResourceClassSimpleName());
     }
 
     @Override
     public void onShutdown() {
         getLog().infoSimTime(this, "Shutdown application");
     }
-
-    @Override
-    public void onAcknowledgementReceived(ReceivedAcknowledgement acknowledgedMessage) {
-    }
-
-    @Override
-    public void onCamBuilding(CamBuilder camBuilder) {
-    }
-
-    @Override
-    public void onMessageTransmitted(V2xMessageTransmission v2xMessageTransmission) {
-    }
-
 }
