+++ bundle/src/assembly/resources/scenarios/Tiergarten/mapping/mapping_config.json
@@ -3,6 +3,7 @@
         {
             "name": "Car",
             "applications": [
+                "org.eclipse.mosaic.app.tutorial.MessageReceivingApp",
                 "org.eclipse.mosaic.app.tutorial.EventSendingApp",
                 "org.eclipse.mosaic.app.tutorial.EventProcessingApp"
             ],
